Update Student.java
I have edited the code at line 4 so that the initial value of totalNumberOfStudent is 0 instead of a 1 as no object of the student class currently exists. I have also added and increment operator inside the constructor to update the value of numberOfStudent each time the constructor is called to create a student object and lastly i have added some comments to document the code.